i was doing commodity markets wrong, and then i fixed it!

+++ seafarers_of_splorr/Data.Game.Island.Market.h
@@ -8,8 +8,8 @@ namespace data::game::island::Market
 	{
 		double supply;
 		double demand;
-		int purchases;
-		int sales;
+		double purchases;
+		double sales;
 	};
 
 	void Write(const common::XY<double>&, int, const MarketData&);
